Maj writers and thread + main

+++ BankProject/src/bankproject/readers/AbstractReader.java
@@ -6,10 +6,11 @@ import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
 
-public abstract class AbstractReader extends Thread { 
-	
+public abstract class AbstractReader extends Thread {
+
 	File file;
 	BufferedReader input = null;
+	Boolean fileexist = false;
 
 	public AbstractReader() {
 
@@ -18,51 +19,55 @@ public abstract class AbstractReader extends Thread {
 	public String getFileInputPath() {
 		return null;
 	}
-	
-	
-	public  String getFileInputPrimaryPath() {
-		
+
+	public String getFileInputPrimaryPath() {
+
 		String fs = System.getProperty("file.separator");
 		String dirPath = System.getProperty("user.dir") + fs + "tmp" + fs + "bank" + fs + "input" + fs;
 		return dirPath;
-		
+
 	}
-	
-	
+
 	protected void readFile() {
-		
+
 		if (!file.exists()) {
 			System.err.println("# Error : \"" + getFileInputPath() + "\" n'existe pas.");
-			System.exit(1);
+			fileexist = false;
+			// System.exit(1);
+		} else {
+			fileexist = true;
 		}
-		
-		try {
-			input = new BufferedReader(new FileReader(file));
-		} catch (FileNotFoundException e) {
-			System.err.println("# Error : impossible to read \"" + file + "\".");
-			System.exit(1);
-		}
-		
-		readSpecificFile();
-		
-		try {
-			input.close();
-		} catch (IOException e) {
-			System.err.println("# Error on closing \"" + file + "\".");
-			System.exit(1);
-		}
-		
-		// Suppression du fichier
-		// TODO Réactiver la suppression quand les tests seront terminés
 
-		 deleteFile(file);
+		if (fileexist) {
+
+			try {
+				input = new BufferedReader(new FileReader(file));
+			} catch (FileNotFoundException e) {
+				System.err.println("# Error : impossible to read \"" + file + "\".");
+				// System.exit(1);
+			}
+
+			readSpecificFile();
+
+			try {
+				input.close();
+			} catch (IOException e) {
+				System.err.println("# Error on closing \"" + file + "\".");
+				System.exit(1);
+			}
+
+			// Suppression du fichier
+			// TODO Réactiver la suppression quand les tests seront terminés
+
+			deleteFile(file);
+		}
 
 	}
 
 	public void readSpecificFile() {
-		
+
 	}
-	
+
 	public void deleteFile(File file) {
 
 		if (!file.delete()) {
@@ -70,18 +75,5 @@ public abstract class AbstractReader extends Thread {
 			System.exit(1);
 		}
 	}
-	
-	
-	
-	
-	
-	
-	
-	
-	
-
-
-
-
 
 }
